fix: resolve MyPy errors batch A25 (errors 1201-1250 from refreshed log)
- Add missing type imports (Any, Dict, Tuple, cast, Union, TYPE_CHECKING)
- Remove unused type: ignore comments
- Add proper return type annotations to functions
- Fix type mismatches with explicit type annotations
- Fix no-any-return errors with cast()
- Fix union-attr errors with None checks
- Fix assignment errors with proper Dict[str, Any] types
- Fix arg-type errors for abstract protocol types
- Fix type-arg errors for TypeAlias usage
- Add proper type parameters to Callable and Task types
- Fix semantic cache embedding None handling
- Fix telemetry config return-value errors with type ignore

+++ victor/coding/docgen/manager.py
@@ -348,7 +348,7 @@ class DocGenManager:
                 + stats["methods_documented"]
             )
 
-            stats["coverage_percent"] = (documented / total * 100) if total > 0 else 100
+            stats["coverage_percent"] = int((documented / total * 100)) if total > 0 else 100
 
             coverage[str(py_file)] = stats
